requirements.txt
# Multi-Agent Content Creation System - Requirements
# =================================================

# Core Framework
crewai>=0.28.0
crewai-tools>=0.1.0

# Web Interface
streamlit>=1.28.0

# Environment Management
python-dotenv>=1.0.0

# LLM Integration
litellm>=1.0.0

# Optional: For advanced features
# openai>=1.0.0  # Uncomment if using OpenAI models
# anthropic>=0.3.0  # Uncomment if using Claude models